สร้าง class ใหม่แล้วเชื่อม class เข้าด้วยกัน
